+++ TractorMod/release-notes.md
@@ -1,6 +1,7 @@
 # Release notes
 ## Upcoming release
 * Updated for Stardew Valley 1.4.
+* Added option to allow summoning a temporary tractor without building a garage first.
 * Improved translations. Thanks to Hesperusrus (updated Russian)!
 
 ## 4.8.4